+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_face.java
@@ -33,13 +33,18 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  * @author Samuel Audet
  */
 @Properties(inherit = {opencv_dnn.class, opencv_objdetect.class, opencv_photo.class, opencv_video.class}, value = {
-    @Platform(include = {"<opencv2/face/predict_collector.hpp>", "<opencv2/face.hpp>", "<opencv2/face/facerec.hpp>"},
+    @Platform(include = {"<opencv2/face/predict_collector.hpp>", "<opencv2/face.hpp>", "<opencv2/face/facerec.hpp>",
+                         "<opencv2/face/facemark.hpp>", "<opencv2/face/facemarkLBF.hpp>", "<opencv2/face/facemarkAAM.hpp>",
+                         "<opencv2/face/face_alignment.hpp>"},
               link = "opencv_face@.3.4", preload = {"opencv_plot@.3.4", "opencv_tracking@.3.4"}),
     @Platform(value = "windows", link = "opencv_face340", preload = {"opencv_plot340", "opencv_tracking340"})},
         target = "org.bytedeco.javacpp.opencv_face")
 public class opencv_face implements InfoMapper {
     public void map(InfoMap infoMap) {
         infoMap.put(new Info("cv::face::BasicFaceRecognizer", "cv::face::EigenFaceRecognizer",
-                             "cv::face::FisherFaceRecognizer", "cv::face::LBPHFaceRecognizer").purify());
+                             "cv::face::FisherFaceRecognizer", "cv::face::LBPHFaceRecognizer",
+                             "cv::face::FacemarkLBF", "cv::face::FacemarkAAM").purify())
+               .put(new Info("cv::face::FN_FaceDetector").cast().valueTypes("Pointer"))
+               .put(new Info("bool (*)(cv::InputArray, cv::OutputArray, void*)").cast().pointerTypes("Pointer"));
     }
 }